requirements.txt
# Core Web Framework
Django==5.2.4

# Optional API & Forms Tools
django-tastypie==0.15.1  # Consider replacing with Django REST Framework
djangorestframework==3.16.0
djangorestframework-xml==2.0.0  # For XML support in DRF
django-filter==25.1
drf-spectacular==0.28.0
django-cors-headers==4.7.0  # For handling CORS in APIs

# HTML and Image Rendering
Pillow

# HTTP & OAuth
requests==2.32.4
requests-oauthlib==1.3.1
oauthlib==3.2.2
urllib3==2.5.0
idna==3.7
charset-normalizer==3.3.2
certifi==2024.7.4
httplib2==0.22.0  # Keep only if you know it’s being used

# Utilities
python-dateutil==2.9.0.post0
six==1.16.0
pyparsing==3.1.4
typing_extensions==4.11.0
packaging==24.0
importlib-metadata==7.1.0
zipp==3.19.1

# Markdown Support
Markdown==3.5.2

# Documentation (optional, for Sphinx)
Sphinx==7.3.7
alabaster==0.7.14
Babel==2.14.0
docutils==0.21.2
imagesize==1.4.1
snowballstemmer==2.2.0
Pygments==2.17.2
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
